Enable non-ZeRO mode (#7515)
Enabled via `stage=0` which corresponds to DDP. 
Remove hardwired path to b16_optimizer.
Enable`torch.autocast` for DDP training
Enable native mixed precision DDP for bfloat16
Update torch.autocast and native mixed precision UTs

<img width="976" height="184" alt="image"
src="https://github.com/user-attachments/assets/92904cdc-e312-46a4-943f-011eb5ab146a"
/>

---------

+++ tests/unit/runtime/test_ds_initialize.py
@@ -229,7 +229,7 @@ class TestOptimizerImplementation(DistributedTest):
         is_supported[(None, 'fp16', None)] = True
         is_supported[(None, 'fp16', 'fp16')] = True
         # BF16 Wrapper
-        is_supported[(None, 'bf16', 'fp32')] = True
+        is_supported[(None, 'bf16', 'bf16')] = True
         is_supported[(None, 'bf16', None)] = True
         # No Wrapper
         is_supported[(None, 'fp32', None)] = True
